# src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Modifisert api basert på mattilsynets åpne data som kan finnes på https://data.norge.no/datasets/288aa74c-e3d3-492e-9ede-e71503b3bfd9",
    "version": "0.0.1",
    "title": "Tilsynsapi"
  },
  "host": "smilefjes.herokuapp.com",
  "basePath": "/",
  "tags": [
    {
      "name": "tilsyn",
      "description": "Hent en liste med tilsyn"
    }
  ],
  "schemes": [
    "https"
  ],
  "paths": {
    "/tilsyn": {
      "get": {
        "tags": [
          "tilsyn"
        ],
        "summary": "Hent en liste med tilsyn",
        "description": "",
        "operationId": "hentTilsynsliste",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "postnummer",
            "in": "query",
            "description": "Filtrer på postnummer",
            "required": false,
            "type": "string",
            "example": "7010"
          },
          {
            "name": "poststed",
            "in": "query",
            "description": "Filtrer på poststed",
            "required": false,
            "type": "string",
            "example": "TRONDHEIM"
          },
          {
            "name": "smilefjes",
            "in": "query",
            "description": "Filtrer på smilefjes",
            "required": false,
            "type": "string",
            "items": {
              "type": "string",
              "enum": [
                "BLID",
                "NØYTRAL",
                "SUR"
              ]
            },
            "example": "BLID"
          }
        ],
        "responses": {
          "200": {
            "description": "En liste med tilsyn",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Tilsyn"
              }
            }

          }
        }
      }
    },
    "/tilsyn/{tilsynsId}": {
      "get": {
        "tags": [
          "tilsyn"
        ],
        "summary": "Hent detaljer om et tilsyn",
        "description": "",
        "operationId": "HentDetaljerOmTilsyn",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "tilsynsId",
            "required": true,
            "type": "string",
            "maximum": 1,
            "description": "Tilsynets id som er gitt i UUID-format",
            "example": "010827c2-3c6b-4095-a654-9cb5f819a50a"
          }
        ],
        "responses": {
          "200": {
            "description": "Tilsynsdetaljer",
            "schema": {
              "type": "object",
              "$ref": "#/definitions/TilsynDetaljer"
            }

          }
        }
      }
    }
  },
  "definitions": {
    "Tilsyn": {
      "type": "object",
      "properties": {
        "tilsynsId": {
          "type": "string",
          "format": "uuid",
          "example": "a2831c9e-76d9-4113-8235-571402be4be4"
        },
        "navn": {
          "type": "string",
          "example": "Super Hero Burger"
        },
        "dato": {
          "type": "string",
          "example": "19062019",
          "description": "Datoen tilsynet ble gjennomført. Har formatet DDMMYYYY"
        },
        "postnummer": {
          "type": "string",
          "minLength": 4,
          "maxLength": 4,
          "example": "7010"
        },
        "smilefjes": {
          "type": "string",
          "description": "Tekstlig representasjon av tilsynets resultat",
          "enum": [
            "BLID",
            "NØYTRAL",
            "SUR"
          ],
          "example": "SUR"
        },
        "poststed": {
          "type": "string",
          "enum": [
            "OSLO",
            "BERGEN",
            "TRONDHEIM"
          ],
          "example": "TRONDHEIM"
        }
      }
    },
    "TilsynDetaljer": {
      "type": "object",
      "properties": {
        "tilsynsId": {
          "type": "string",
          "format": "uuid",
          "example": "a2831c9e-76d9-4113-8235-571402be4be4"
        },
        "navn": {
          "type": "string",
          "example": "Super Hero Burger"
        },
        "dato": {
          "type": "string",
          "example": "19062019",
          "description": "Datoen tilsynet ble gjennomført. Har formatet DDMMYYYY"
        },
        "postnummer": {
          "type": "string",
          "minLength": 4,
          "maxLength": 4,
          "example": "7010"
        },
        "smilefjes": {
          "type": "string",
          "description": "Tekstlig representasjon av tilsynets resultat",
          "enum": [
            "BLID",
            "NØYTRAL",
            "SUR"
          ],
          "example": "SUR"
        },
        "poststed": {
          "type": "string",
          "enum": [
            "OSLO",
            "BERGEN",
            "TRONDHEIM"
          ],
          "example": "TRONDHEIM"
        },
        "tilsynshistorikkId": {
          "type": "string",
          "example": "Z1212211006466393252ZMNMZ",
          "description": "En id som er lik for alle tilsyn gjennomført på samme restaurant"
        },
        "rutinerOgLedelse": {
          "type": "integer",
          "example": "2",
          "minimum": 0,
          "maximum": 5,
          "description": "Et heltall som beskriver hvilken score matstedet har fått innen kategorien Rutiner og ledelse"
        },
        "lokalerOgUtstyr": {
          "type": "integer",
          "example": "2",
          "minimum": 0,
          "maximum": 5,
          "description": "Et heltall som beskriver hvilken score matstedet har fått innen kategorien Lokaler og utstyr"
        },
        "mathåndteringOgTilberedning": {
          "type": "integer",
          "example": "2",
          "minimum": 0,
          "maximum": 5,
          "description": "Et heltall som beskriver hvilken score matstedet har fått innen kategorien Mathåndtering og tilberedning"
        },
        "merkingOgSporbarhet": {
          "type": "integer",
          "example": "2",
          "minimum": 0,
          "maximum": 5,
          "description": "Et heltall som beskriver hvilken score matstedet har fått innen kategorien Merking og sporbarhet"
        },
        "adresse": {
          "type": "string",
          "example": "Kaia 9",
          "description": "Adressen til spistedet hvor tilsynet ble gjennomført."
        }
      }
    }
  }
}